third_party/WebKit/LayoutTests/fast/css/usecounter-pseudo-any.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style id="style"></style>
<script>
'use strict';

test(() => {
    var CSSSelectorPseudoAny = 1250; // From UseCounter.h
    var style = document.getElementById('style');
    assert_false(internals.isUseCounted(document, CSSSelectorPseudoAny));
	style.textContent = ":-webkit-any(.a1, .a2) { color: blue; }";
    assert_true(internals.isUseCounted(document, CSSSelectorPseudoAny));
}, 'webkit-any is use counted.');
</script>
