third_party/WebKit/LayoutTests/platform/win/fast/css/usecounter-pseudo-any-expected.txt

This is a testharness.js-based test.
PASS webkit-any is use counted. 
Harness: the test ran to completion.
